tools/requirements.txt
boto3>=1.34
pyarrow>=15.0
python-dotenv>=1.0
